% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axCorrgram.R
\name{axCorrgram}
\alias{axCorrgram}
\title{Corrgram-like correlation table plot}
\usage{
axCorrgram(data, fontsize = 3)
}
\arguments{
\item{data}{Matrix / DataFrame with items in columns. For
parallelizedBy =="IAT", data must be a logfile prepared for the cleanIAT()
function of the IAT package.}

\item{fontsize}{Base font size of the correlations in cex}
}
\value{
Returns a data frame containing the parcels.
}
\description{
In the upper right triangle the correlations are plotted as numbers, and in the
lower left triangle a scatterplot with smoothed line is plotted.
In the diagonal histogramms are plotted
}
\examples{
# Axel: fix me


}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
