% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axFormatTableDat.R
\name{axFormatTableDat}
\alias{axFormatTableDat}
\title{Get the (apa formatted) output of a linear (moderated) regression in a variable list}
\usage{
axFormatTableDat(
  data,
  pred,
  crit,
  cov = NULL,
  mod = NULL,
  symMargSig = "+",
  symSig = "*",
  sorting = "",
  formatted = "apa"
)
}
\arguments{
\item{data}{A data frame containing the numeric columns to be used in the
regression.}

\item{pred}{Character with the column name which should be used as
predictors of the linear regression}

\item{crit}{Character cointaining the column name of the criteria}

\item{cov}{Character vector with the column names which should be used as
covariates in the linear regression}

\item{mod}{Character with the column name which sould be used as a moderator
(pred * mod) in the linear regression}

\item{symMargSig}{Symbol to be used as indicator of marginal significance in
formatted (apa) output. Default is dagger}

\item{symSig}{Symbol to be used as indicator of significance in
formatted (apa) output. Default is *}

\item{sorting}{Character vector containing the sequence of variables which
should be used in the output}

\item{formatted}{Should the variable list contain apa formatted output
("apa", e.g., rounded, no leading zero, stars and daggers of significance)
or unformatted ("raw") output. Default is "apa".}
}
\value{
Returns a list with variables of the regression output, either apa
formatted or raw to be used for further computations or table output.
}
\description{
Get the (apa formatted) output of a (moderated) regression in a variable list
to be used in further computations or table arrangements.
}
\examples{
axFormatTableDat(axBoost::df1, pred = "IndMeasure", crit = "AutoBehav",
cov = c("DirMeasure"), mod = "mod", formatted = "apa")


}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
