% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDD.R
\name{drawDD}
\alias{drawDD}
\title{Draw a double dissociation based on regression coefficients}
\usage{
drawDD(
  data,
  predUpper,
  predLower,
  critUpper,
  critLower,
  moderator = "",
  moderatorSD = c(-1, 1),
  modBothPaths = FALSE,
  printReg = FALSE,
  title = ""
)
}
\arguments{
\item{data}{A data frame containing the numeric columns predUpper,
predLower, critUpper, critLower, and optionally moderator.}

\item{predUpper}{String with the name of the Predictor of the upper double dissociation path.}

\item{predLower}{String with the name of the Predictor of the lower double dissociation path.}

\item{critUpper}{String with the name of the Criterion of the upper double dissociation path.}

\item{critLower}{String with the name of the Criterion of the lower double dissociation path.}

\item{moderator}{String with the name of the Moderator of the upper and lower direct path from
predictor to criterion}

\item{moderatorSD}{Defines the standard deviations for which the slope
lines are plotted. Default is -1 and 1. For example, for experimetal group
/ control group design use c(0,1).}

\item{modBothPaths}{If TRUE, a moderator effect is computed for both paths}

\item{printReg}{Print the summary () of the regression coefficients.}

\item{title}{Title of plot}
}
\value{
Returns a plot with a (moderated) double dissociation.
}
\description{
Draws a double dissociation based on regression coefficients.
}
\details{
Assuming a double dissociation, the function computes two regressions:

critUpper ~ predUpper + predLower

critLower ~ predLower + predUpper

If a moderator is given, the regressions are computed as:

critUpper ~ predUpper * moderator + predLower

critLower ~ predLower * moderator + predUpper

If modBothPaths = TRUE, the regressions are computed as:

critUpper ~ predUpper * moderator + predLower * moderator

critLower ~ predLower * moderator + predUpper * moderator

All variables are scaled before computing the regressions to cope with
multicolinearity in the moderated case.
}
\examples{
# Axel: fix me

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
