% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillVec.R
\name{fillVec}
\alias{fillVec}
\title{Add missing elements to a vector of consecutive numbers}
\usage{
fillVec(data, replValue = NA, verbose = FALSE)
}
\arguments{
\item{data}{A numeric vector which should contain consecutive numbers,
but isn't.}

\item{replValue}{A numeric atomic vector which is used for replacement
(default is NA).}

\item{verbose}{Logical value, if verbose output should be given (default is
FALSE).}
}
\value{
Returns a numeric vector in which missing consecutive values are
replaced by replValue (e.g., NA).
}
\description{
Fill a vector, which should be of consecutive numbers, but due to missings
is not containing only consecutive numbers (e.g., the residuals of a lm())
with values (default is NA).
}
\examples{
data = c(1,2,3,4,5,6,7,8,10,11,12,13,16,17,18,20,25,30)
dataMod <- fillVec(data,verbose = TRUE)

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
