% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modRegGraph.R
\name{modRegGraph}
\alias{modRegGraph}
\title{Plot a moderated regression interaction diagram and return graphical coordinates}
\usage{
modRegGraph(
  betas,
  pred,
  mod,
  crit,
  modRangeL = -1,
  modRangeH = 1,
  title = "",
  plot = TRUE
)
}
\arguments{
\item{betas}{Coefficients (Betas) of the moderated regression}

\item{pred}{String with the name of the predictor}

\item{mod}{String with the name of the moderator}

\item{crit}{String with the name of the criterium}

\item{modRangeL}{Lower end of the moderator to be plottet (default: -1 SD)}

\item{modRangeH}{Upper end of the moderator to be plottet (default: 1 SD)}

\item{title}{Title of the plot}

\item{plot}{Logical value if a plot should be printed (default is TRUE)}
}
\value{
Named vector with the four coordinates (pred_H_mod_H, pred_H_mod_L,
pred_L_mod_H, pred_L_mod_L) on the y-axis for predictor at -1/1 SD
and moderator at modRangeL/modRangeH SD
}
\description{
Function plots a moderated regression interaction diagramm and returns the
coordinates of the points pred_H_mod_H pred_H_mod_L pred_L_mod_H pred_L_mod_L
to be used elswhere (e.g., in another plot).
}
\examples{
# Axel: fix me

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
