% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcelMe.R
\name{parcelMe}
\alias{parcelMe}
\title{Parcels a dataframe with item columns}
\usage{
parcelMe(data, nParcels = 3, parcelName = "parcel",
  parallelizedBy = "loadings")
}
\arguments{
\item{data}{Matrix / DataFrame with items in columns. For
parallelizedBy =="IAT", data must be a logfile prepared for the cleanIAT()
function of the IAT package.}

\item{nParcels}{Number of parcels to build. Default is 3.}

\item{parcelName}{Name stem of parcel names to which the parcelnumber
will be added. Default is "parcel".}

\item{parallelizedBy}{Method to build parallel parcels. Can be "loadings",
"discrimination" or "IAT". Factor loadings are computed by a PAF with oblique
rotation. Default is "loadings". For "IAT" the data must be a logfile prepeared
for the use of the cleanIAT() function of the IAT package.}
}
\value{
Returns a data frame containing the parcels.
}
\description{
The function accepts a data frame containing items and generates parcels,
e.g. for Structural Equation Modelling, according to the parameters
specified below.
}
\examples{
# Axel: fix me

}
\author{
Axel Zinkernagel \email{zinkernagel@uni-landau.de}
}
